Make blog title and description more legible on light backgrounds
Adds a light shadow to blog title and description to give slightly more contrast between text and background when a cover image is used with a predominently light colour. This is not a complete fix, because heavy shadows do not lend to the aesthetic of Casper - however it does go some way to making a small improvement to legibility in most cases.

Related: https://github.com/TryGhost/Ghost/issues/1768